Fix bug

+++ include/RoutePlannerContext.h
@@ -99,7 +99,7 @@ namespace OsmAnd {
         class OSMAND_CORE_API RoutingSubsectionContext
         {
         private:
-            uint64_t _mixedLoadsCounter;
+            uint32_t _mixedLoadsCounter;
         protected:
             RoutingSubsectionContext(RoutePlannerContext* owner, const std::shared_ptr<ObfReader>& origin, const std::shared_ptr<ObfRoutingSection::Subsection>& subsection);
